refactor : 가장 멀리 이동한 자동차의 위치 구하는 메소드 이동
WinningRule.java -> Cars.java
메소드 이동에 따라서 WinningRule에 maxPosition field 추가
